Implement twin() Method in board.java
twin() makes a copy of the current board and then generates uniform random
indicies to swap any pair of elements in the twin board. The empty tile
(i.e where board[i][j] == 0) is not allowed to be swapped, so if the random
indicies land on this case the indices must be recalculated.